Revert "[CFM_57] API to Get All Fridges and Filter by Tag (#17)"
This reverts commit 6460dd352199ba26ec98e4e0b92b23a26bfa457c.
