+++ CommunityFridgeMapApi/functions/dev/load_fridge_data/app.py
@@ -19,7 +19,6 @@ FRIDGE_DATA = [
             "state": "NY",
             "city": "Brooklyn",
         },
-        "tags": ["tag1", "tag3"],
         "maintainer": {"instagram": "https://www.instagram.com/thefriendlyfridge/"},
     },
     {
@@ -32,7 +31,6 @@ FRIDGE_DATA = [
             "state": "NY",
             "city": "Brooklyn",
         },
-        "tags": ["tag1", "tag4"],
         "maintainer": {"instagram": "https://www.instagram.com/2fish5loavesfridge/"},
     },
 ]
